исправил багу

+++ hl_mai_06_cache/database/cache.h
@@ -13,7 +13,7 @@ namespace database
         public:
             static Cache get();
             void put(long id, const std::string& val);
-            bool get(int id, std::string& val);
+            bool get(long id, std::string& val);
             size_t size();
             void remove(long id);
             void remove_all();
